Do not set BASH_ENV
I think had wrongly assumed that this is used to source the startup file
for interactive shells. (As in fact $ENV is sourced by Posix shells.)
That's not the case: $BASH_ENV is sourced even for #!/bin/bash scripts.
Feh. I caught this when the "cd" command didn't work in a script (it
printed its directory argument to stdout and used CDPATH to expand it.)